feat: more diverse stages in app dockefile, 3 compose files - prod, test, dev, unit-tests now cover whole app/domain layer for 100%. Todo: integration tests

+++ services/api/app/infrastructure/security/auth_strategies.py
@@ -14,53 +14,82 @@ from app.presentation.schemas import TokenResponse
 import typing as t
 import jwt, uuid, datetime as dt
 
-REFRESH_SECRET = Config.REFRESH_SECRET
-JWT_SECRET = Config.JWT_SECRET
-ACCESS_EXPIRES_MINS = Config.ACCESS_TOKEN_EXPIRE_MINUTES
-REFRESH_EXPIRES_HOURS = Config.REFRESH_TOKEN_EXPIRE_HOURS
-ALGORITHM = Config.ALGORITHM
 
 class StatefulOAuthStrategy(iapp.IAuthStrategy, iapp.ITokenMixin, iapp.IPasswordMixin, iapp.ILoginLogoutMixin):
 
-    def __init__(self, session_repo: SessionRepository, user_repo: repos.IUserRepository, password_hasher: domsvc.IPasswordHasher):
+    def __init__(
+        self,
+        session_repo: SessionRepository,
+        user_repo: repos.IUserRepository,
+        password_hasher: domsvc.IPasswordHasher,
+        *,
+        refresh_secret: t.Optional[str] = None,
+        jwt_secret: t.Optional[str] = None,
+        access_expires_mins: t.Optional[int] = None,
+        refresh_expires_hours: t.Optional[int] = None,
+        algorithm: t.Optional[str] = None,
+    ):
+        
         self.session_repo = session_repo
         self.user_repo = user_repo
         self._hasher = password_hasher
 
-    def __exctract_token_data(self, token: str, refresh:bool=False):
-        secret = REFRESH_SECRET if refresh else JWT_SECRET
+        self.refresh_secret = refresh_secret or Config.REFRESH_SECRET
+        self.jwt_secret = jwt_secret or Config.JWT_SECRET
+        self.access_expires_mins = (
+            access_expires_mins or Config.ACCESS_TOKEN_EXPIRE_MINUTES
+        )
+        self.refresh_expires_hours = (
+            refresh_expires_hours or Config.REFRESH_TOKEN_EXPIRE_HOURS
+        )
+        self.algorithm = algorithm or Config.ALGORITHM
+
+    def __exctract_token_data(self, token: str, refresh: bool = False):
+        secret = self.refresh_secret if refresh else self.jwt_secret
         try:
-            data = jwt.decode(token, secret, algorithms=[ALGORITHM])
+            data = jwt.decode(token, secret, algorithms=[self.algorithm])
             return data
-        except jwt.InvalidTokenError as e:
-            raise appexc.CredentialsException() from e
         except jwt.ExpiredSignatureError as e:
             raise appexc.TokenExpiredException() from e
+        except jwt.InvalidTokenError as e:
+            raise appexc.CredentialsException() from e
 
 
-    @staticmethod
-    def __create_token(payload: dict, expires_delta: dt.timedelta, refresh: bool = False) -> str:
-        secret = REFRESH_SECRET if refresh else JWT_SECRET
+    def __create_token(self, payload: dict, expires_delta: dt.timedelta, refresh: bool = False) -> tuple[str, float]:
+        secret = self.refresh_secret if refresh else self.jwt_secret
         expiration_time = (dt.datetime.now(dt.timezone.utc) + expires_delta).timestamp()
-        encoded_jwt = jwt.encode(payload | {"exp": expiration_time}, secret, algorithm=Config.ALGORITHM)
+        encoded_jwt = jwt.encode(
+            payload | {"exp": expiration_time}, secret, algorithm=self.algorithm
+        )
         return encoded_jwt, expiration_time
 
-    @staticmethod
-    def __create_a_pair_of_tokens(payload: dict) -> TokenResponse:
-        access_token_expires = dt.timedelta(minutes=ACCESS_EXPIRES_MINS)
-        refresh_token_expires = dt.timedelta(days=REFRESH_EXPIRES_HOURS)
+    def __create_a_pair_of_tokens(self, payload: dict) -> TokenResponse:
+        access_token_expires = dt.timedelta(minutes=self.access_expires_mins)
+        refresh_token_expires = dt.timedelta(hours=self.refresh_expires_hours)
 
-        access_token, access_expires = StatefulOAuthStrategy.__create_token(payload=payload, expires_delta=access_token_expires)
-        refresh_token, refresh_expires = StatefulOAuthStrategy.__create_token(payload=payload, expires_delta=refresh_token_expires, refresh=True)
-        return TokenResponse(access_token=access_token, refresh_token=refresh_token, access_expires=access_expires, refresh_expires=refresh_expires)
+        access_token, access_expires = self.__create_token(
+            payload=payload, expires_delta=access_token_expires
+        )
+        refresh_token, refresh_expires = self.__create_token(
+            payload=payload, expires_delta=refresh_token_expires, refresh=True
+        )
+        return TokenResponse(
+            access_token=access_token,
+            refresh_token=refresh_token,
+            access_expires=access_expires,
+            refresh_expires=refresh_expires,
+        )
 
     @property
     def hasher(self) -> domsvc.IPasswordHasher:
         return self._hasher
 
     async def login(self, credentials: dict) -> TokenResponse:
-        username = credentials['username']
-        password = credentials['password']
+        username = credentials.get('username')
+        password = credentials.get('password')
+
+        if not (username and password):
+            raise appexc.CredentialsException("Field missing! Both username and password must be provided.")
         
         user = await self.user_repo.get_by_username(username)
         if not user:
@@ -70,25 +99,25 @@ class StatefulOAuthStrategy(iapp.IAuthStrategy, iapp.ITokenMixin, iapp.IPassword
             raise appexc.CredentialsException("Bad password given for this username!")
         
         session_id = str(uuid.uuid4())
-        tokens = self.__create_a_pair_of_tokens(dict(session_id = session_id))
+        tokens = self.__create_a_pair_of_tokens(dict(session_id=session_id))
         session = mapp.RotatingTokenSession(
             id = session_id,
             user_id = user.id,
             roles = [user.role],
             refresh_token=tokens.refresh_token
         )
-        
-        
-        await self.session_repo.create(session, ttl=ACCESS_EXPIRES_MINS*60)
+        await self.session_repo.create(session, ttl=self.access_expires_mins * 60)
         return tokens
 
     async def logout(self, credentials: dict) -> None:
-        token = credentials['token']
+        token = credentials.get('token')
+        if not token:
+            raise appexc.CredentialsException("Token is missing. Please provide a valid access token for this operation.")
         data = self.__exctract_token_data(token=token)
         await self.session_repo.delete(data["session_id"])
     
     async def authenticate(self, credentials: dict) -> mdom.User:
-        token = credentials["token"]
+        token = credentials.get('token')
         if not token:
             raise appexc.CredentialsException("Token is missing")
         data = self.__exctract_token_data(token=token, refresh=False)
@@ -108,12 +137,12 @@ class StatefulOAuthStrategy(iapp.IAuthStrategy, iapp.ITokenMixin, iapp.IPassword
         if not session:
             raise appexc.LoggedOutException("Token is valid, yet session does not exist!")
         
-        if session.refresh_token!=refresh_token:
-            raise appexc.TokenExpiredException("This has been rotated already!")
-        
-        tokens = self.__create_a_pair_of_tokens(dict(session_id = session.id))
+        if session.refresh_token != refresh_token:
+            raise appexc.TokenExpiredException("This token has been rotated already - it not valid anymore!")
+
+        tokens = self.__create_a_pair_of_tokens(dict(session_id=session.id))
         session.refresh_token = tokens.refresh_token
-        await self.session_repo.create(session, REFRESH_EXPIRES_HOURS * 3600)
+        await self.session_repo.create(session, self.refresh_expires_hours * 3600)
         return tokens
